Adding tracker test
